AI_NYC_Files
Included is the coursework requirements specification, our submission (in a .ZIP file with the dataset included), the coursework report of our group (written by me) including: what is our dataset and problem domain, the problem domain (whether we used classification or regression), how we managed data cleaning and handling missing data, techniques used for our dataset, data encoding using one hot encoding and categorical encoding, model outputs, and problems/difficulties with our dataset and the conclusion of our project. 

This was written by me primarily with support from my team, and included in this project is my individual reflection of the project.